build: rename summary to gauge
